Move loading state config to sup module and refactor doc tests
